extracted out getDefaultNamespace method
